Add build rules to modify ICUConfig.properties to set dataPath.
The ICUConfig.properties in the source tree doesn't set any value for
dataPath, which is how it should be when loading ICU4J data from the
JAR. But for the BUILD_STATIC_JAVA_LIBRARY build rule, which builds the
library to be executed on an actual device, dataPath should be set to
the location of the ICU4C data files in the Android device file system,
so that ICU4J prefers to load ICU4C data files instead of using its own.

These new build rules will dynamically generate an ICUConfig.properties
file with the dataPath value appended and include this file instead of
the one from the source tree in the generated JAR.

This change also removes the BUILD_STATIC_JAVA_LIBRARY dependency on the
ICU4J data JARs, which are no longer needed when loading ICU4C data.

Change-Id: I63a254a640f1831677ca4c6378246cec09ef5edb