Android patch: Work-around for ClassNotFoundException.
Android throws ClassNotFoundException when trying to deserialize objects
containing references to classes that don't exist. It has not yet been
determined whether this is working as intended or whether this is a bug,
so in the meantime this local patch for the Android copy of ICU4J makes
the code expect it (as this is how Android actually works today,
intentional or not).

Bug: 19109180
Change-Id: Ibd31b41e89de4286aa0b4fe431c441e84d98b0eb