Remove header from LOCAL_SRC_FILES
The build system does not use headers in LOCAL_SRC_FILES. To prevent
typos, it's becoming a warning to add something to LOCAL_SRC_FILES that
is not used.

Also removes the Android.mk from LOCAL_ADDITIONAL_DEPENDENCIES. Our make
setup now automatically adds it, and kati no longer needs it.

And explicitly set optional_android_logging_includes/libraries so that
definitions from elsewhere in the tree will not be used.

Change-Id: I245d1d112a3e4ae19ee5ef5ccecbb49d72b03376