Enable tests to be run when ICU4J is on the bootclasspath
Applies the patch attached to
http://bugs.icu-project.org/trac/ticket/12227 which supplies the
ClassLoader for a test class to the code that retrieves the data
rather than relying on its default behaviour which is to use the
ClassLoader for one of the main library classes which is in the
bootclassloader.

Also repackages the code to make sure that the changes are
reflected in the repackaged tests that are run by CTS.

This is needed to run the tests on the repackage code embedded
within Android.

Bug: 22023363
Change-Id: I7f1c8a741e8bd271b0f355d7985a72dc63e8c3a5