Repackage serialized data and fix test so it can find it
Fortunately, the original text "org.ibm" or "org/ibm" and the
replacement text "android" used in repackaging are both the same
length so it is possible to use sed to do a simple search and
replace to repackage the binary serialized data.

This repackages the serialized data and patches the affected
test so that it can find the data in a JAR without entries for
directories.

The upstream bug tracking the change to the CompatibilityTest
is http://bugs.icu-project.org/trac/attachment/ticket/12386/

Bug: 27310873
Change-Id: I2d45b983cce51c66aabe7e5202159a02c0600e29