Android patch: CLDR Ticket #9495: Gregorian skeleton patterns for yMMMM
This change was introduced in Android for the N release.

This is a temporary workaround until the issue has been fixed upstream:

http://unicode.org/cldr/trac/ticket/9495

This change contains fixes for the yMMMM skeleton for some languages.

The yMMMM skeleton is used in the Android standard date picker widget to
obtain a pattern to use when displaying the Month and Year at the top.

CLDR / ICU 56 introduced incorrect patterns for some languages. The
languages changed here are ones that appear to (logically) order "M y"
based on surrounding patterns. Other languages like si, ne, ml, ky
logically order "y M" and have been left. Other skeletons have not been
changed.

Bug: 28766642
Change-Id: I74cc2a2e3fcc67df71ce69628178e057ff37e98e