Android patch: Fix deadlock between java.util and ICU TimeZone.
There was a deadlock between java.util.TimeZone.setDefault() and
android.icu.util.TimeZone.getDefault():

- the former would lock on j.u.TimeZone.class and the call
  the synchronized a.i.u.TimeZone.clearCachedDefault()
- the latter would synchronize on a.i.u.TimeZone and then call
  the synchronized j.u.TimeZone.getDefault

To ensure every path synchronizes on objects in the same order, this
change adds an explicit synchronize block on j.u.TimeZone in
a.i.u.TimeZone.getDefault().

TimeZoneTest.testSetDefaultDeadlock() is a probabilistic test that was
able to reproduce the problem roughly 8 times out of 10.

This change was introduced in Android after the N release:

https://android.googlesource.com/platform/external/icu/+/15c4b74

Change-Id: If4bcca81ac160e8473ec71bd445c73014b4953e2