Android patch: Fix race condition in ICU TimeZone.getDefault()
TimeZone.getDefault() had a race condition where it could fail
with a NullPointerException if called concurrently with
clearCachedDefault().

This fixes the race condition by always copying the defaultZone value
into a local variable and returning that, to ensure that concurrent sets
to defaultZone don't lead to a NullPointerException on the
cloneAsThawed() call.

This change was introduced in Android after the N release:

https://android.googlesource.com/platform/external/icu/+/b6127dc

Change-Id: I84b344999e848b0d22b2f3ff6fd55fe511141045