Fix race condition in ICU TimeZone.getDefault()
TimeZone.getDefault() had a race condition where it could fail
with a NullPointerException if called concurrently with
clearCachedDefault().

This fixes the race condition by always copying the defaultZone value
into a local variable and returning that, to ensure that concurrent sets
to defaultZone don't lead to a NullPointerException on the
cloneAsThawed() call.

Bug: 30979219
Change-Id: I0bcef8561be8cb6040c6559111a949806b2ce7b7
Test: vogar libcore/luni/src/test/java/libcore/java/util/TimeZoneTest.java