Remove support for skipping execution
CTSv2 generates a list of tests that will be run by running the
tests in a special skip execution mode. Previously, that
required specific support within the Runners that ran the tests.
Following on from previous change in this bug that is no longer
necessary as it generates the list from the Description returned
by the Runners.

Removes all references to AndroidRunnerParams and so the dummy
version provided to allow the tests to be run on the host is no
longer necessary.

Also removes the unused IcuTestRunnerBuilder.

Bug: 30801147
Test: Ran CtsIcuTestCases and ran the tests directly on the host
Change-Id: Ieecd817f40179c0baa9fb0b98f760d0019478a32