Android patch: Be more tolerant about last locale name.
ULocaleTest.TestGetAvailable expected the last available locale to be
zu_ZA. This is correct with the data we ship. Adding additional locales
after this locale would break the test (adding locales that
alphabetically sort before that locale would not be counted as an
error).

Verify that the last locale starts with "z" to ensure we at least don't
lose locales.

This change was introduced in Android for the O release:

https://android.googlesource.com/platform/external/icu/+/82d33ed

(cherry picked from commit 244df5a9e49eb3e0457ee00034a9dfd960057574)

Change-Id: I1726134cf389ca173f51b1ee27a768ca01b03fd9