Be more tolerant about last locale name.
ULocaleTest.TestGetAvailable expected the last available locale to be
zu_ZA. This is correct with the data we ship. Adding additional locales
after this locale would break the test (adding locales that
alphabetically sort before that locale would not be counted as an
error).

Verify that the last locale starts with "z" to ensure we at least don't
lose locales.

Bug: 31841293
Test: Ran ULocaleTest
Change-Id: I3ce11221403da872aceb7169bfe96e721665f53b