Android patch: Add/expose APIs needed by java.time.
Add and expose a small number of methods that are needed by
java.time. They tend to expose functionality that most end user code
will not need, but allows access to ICU internal data or functionality
that should be shared between ICU and java.time.

- Add method to Calendar to get localized pattern string.
- Add constructor to DateFormatSymbols that takes the calendar type as
  string.
- Add method to DateFormatSymbols to get Narrow Eras

This change was introduced in Android for the O release:

https://android.googlesource.com/platform/external/icu/+/22b47ef

Change-Id: Ic91b4a969ab39a94292d51c17c972aebdebb3ed2