DO NOT MERGE : Remove test for TimeZone.getWindowsID("Australia/Eucla")
A couple of OEMs have picked up the windowsZone.txt changes from:
http://bugs.icu-project.org/trac/changeset/39391
or
http://bugs.icu-project.org/trac/changeset/39460

...when updating timezone rules data to 2016g / 2016h. This isn't needed
to update rules but this commit relaxes the ICU test that
fails when they take that change. This allow devices that have taken the
windowsZones.txt change as well as those that have not to pass CTS.
The behavior is effectively left "undefined" for N.

The ability to map this specific IANA timezone ID to a
Windows ID is unlikely to affect actual app compatibility. The
windowsZones.txt change will affect the behavior of some
android.icu.util.TimeZone methods related to Windows IDs.

Bug: 32512035
Bug: 32534123
Test: Ran CtsIcuTestCases without the windowsZone.txt change
Change-Id: If9b83ea359914a4e9d47b03ba77ee4bb5e407b4b