Bug work-around: Delete android.icu.text.DigitListTest
ICU4J 58 has one test, DigitListTest, which depends on being able to
access the non-public class DigitList by being declared in the same
package.

This does not work in CTS which runs the test against the actual ICU4J
library included in libcore, from where the class loader won't allow the
test to load the non-public class regardless of them being declared in
the same package.

Future ICU4J releases are expected to have more tests like this, for
which some kind of better solution will be needed, but for the time
being it is sufficient to just remove DigitListTest while doing the
source code transform so that CTS won't try loading it.

Bug: 33448125
Test: CtsIcuTestCases
Change-Id: I376d9405db7ea92f8c75246eada7b4b236a1fb20