Bugfix: Hide @discouraged annotations referring to hidden methods.
The bug that truncated @discouraged annotations has until now hidden the
problem that there are @discouraged annotations in DecimalFormatSymbols
referring to new methods introduced in ICU 58 which are still hidden in
Android. With that bug fixed they would now become visible in their
entirety, but make no sense in Android.

Therefore, replacements.txt also needs to be updated to remove these
annotations from the Javadoc of the Android transformed source code.

Test: generate_android_icu4j.sh
Change-Id: I6f90c0ea2b41f77a984e85079d498fe6563acdb0