Add sharding annotations to android_icu4j test classes.
Some tests were using too much heap, causing excessive garbage
collection and subsequent timeouts. This change adds an annotation
those specifically, so that they can be run in a separate process,
using their own heaps, so avoiding the timeouts.

Some of these tests were suppressed in the past. The suppressions were
undone when ICU 59 landed, which most likely caused problems on
low-memory devices (b/62374714, whose fix was reverted when ICU 59
landed, and perhaps b/37642683). This change (along with the
AndroidTest.xml change to actually apply the sharding) fixes those
problems. This was tested on fugu on an oc build.

Bug: 62512577
Test: cts-tradefed run cts -m CtsIcuTestCases --primary-abi-only
Change-Id: I013ea26728120bbe3238a7a2dd165606ccad243d