Make updateicudata.py skip JAR files with identical contents.
JAR files are Zip archives and such things as timestamps inside of the
archives will cause the files to always be different each time they're
generated, even when the contents of the archives are identical.

In order to not unnecessarily fill up the change log with commits of new
JAR files with contents identical to the old ones, updateicudata.py is
updated to compare the contents of newly generated JAR files with the
already existing files, ignoring them if they're identical.

This will ensure that each new commit actually contains differences,
making such things as git bisect much more useful with the ICU data.

Test: updateicudata.py
Change-Id: Ia83e892b5c5a13b15bf562c5fbf866fe9f30687e