Fix RelativeDateTimeFormatter apostrophe bug.
Fix a bug ICU RelativeDateTimeFormatter.combineDateAndTime which led to
single-quotes that were meant to quote characters appearing in the
output.

The relevant code uses SimpleFormatterImpl for performance reasons
previously, but that class does not consider ' to be a quoting character
and just copies them into the output. In this patch we use MessageFormat
to handle the pattern instead which handles the single quote as desired.

This is a short-term local workaround until the proper fix has been
found, which may depend on a change in the CLDR specification.

See http://bugs.icu-project.org/trac/ticket/13295 and
http://unicode.org/cldr/trac/ticket/10321 for context and a potential
long-term fix.

Bug: 63745717
Test: CstIcuTestCases
Test: libcore.icu.RelativeDateTimeFormatterTest
Change-Id: I5e722a30145cde2ac4a6670b69d5bbb56e6c5019