Android patch: Copy DecimalFormat_ICU58 to Android specific path.
Create a copy of DecimalFormat_ICU58 for use in java.text.DecimalFormat.
This also requires a copy of DigitList (as DigitList_Android).

This also include applying a patch from commit
b788e2aa40c1e4d4c73adc8149d9d82c19ee1190 to this copy.

This change was introduced in Android for ICU 59:

https://android.googlesource.com/platform/external/icu/+/ea5ab8c

Change-Id: Icf6a0628016e1ca076660bca5623e8231649f01d