Drop ICU test classes in libcore packages.
Previously [1] [2], the following code had merely skipped the test
logic that relied on DecimalFormat_ICU58 at test runtime.

This is no longer sufficient because OpenJDK 9 based builds require
the test code to be in a different package from the code under test
(as long as --patch-module isn't used). Therefore, this CL removes
the three files DecimalFormat_ICU58.java, DigitList{,Test}.java that
previously were in the same package as code under test.
Since DecimalFormat_ICU58 no longer exists, this CL comments out the
references to that class in from NumberFormat{,DataDriven}Test where
previously that logic had only been skipped at test runtime.

This CL also deletes unused codegen logic to make the (now deleted)
class DigitList public.

This CL has a similar effect, and for the same purpose, as CL [3]
but it does so through a local patch to ICU code as opposed to
changes to the codegen.

[1] http://r.android.com/455118 commit 495cb271e305cfb399d463f32210a371198f0abf
[2] http://r.android.com/432265 commit 7d09282b14753890f8b1fe594552c7877ff83f6a
[3] http://r.android.com/539837 commit 430456d6e8fd3a9734218d3072cdba932ec669d2

Test: Treehugger
Bug: 68696705
Change-Id: I8d2a061db01d40ce030df55bd533a9a1b815202f