Remove broken @see tags from ICU javadoc.
This makes it possible to treat broken @see tags as errors. Note that
some are broken because it refers to something not present on Android,
some because it refers to something non-public.

This can be reverted later if documentation is generated from metalava
and metalava is fixed to drop these programmatically, but that's not
the case yet.

Bug: 80570421
Test: make docs, with doclava patched to treat these broken tags as errors
Change-Id: Iba0c0b168fadcd8be48539f76db4a4966da78380