Android patch: Ticket #20116: Remove unnecessary @Deprecated annotation in Edits.Iterator
ICU ticket: https://unicode-org.atlassian.net/browse/ICU-20116

Edits$Iterator.toString is marked both @internal and @deprecated in ICU.
When we try to expose the Edits API in Android, the annotations confuse
our build system, and shows up as a deprecated API in current.txt.

Our code generation has treated @internal API in ICU as hidden API in Android.
Remove @Deprecated to avoid an API entry in current.txt.

When I ran this on my local machine, the API doc and stub artifacts is not affected
by this change, i.e. the method is hidden in API doc and stub before and
after this change.

Note: I believe Android patch is sightly better an additional step in code generation,
because upstream should eventually remove the @deprecated annotation,
If it's the case, this change should become a cherry-pick in the future.

Bug: 113652645
Test: The toString method is hidden in API doc, stubs, and current.txt
Change-Id: I943e8a5ba7861725f85bb42173c990bef753fcc8