Fix android-icu4j-test compilation for OpenJDK 9.
Commit 6e18e29e64773218e1effa32c3de3dd1c794a836 accidentally
applied the patch_module: java.base attribute to both
host and device versions of android-icu4j-tests. This
broke the EXPERIMENTAL_USE_OPENJDK9=true based build
because it caused the ICU sources to be patched into
OpenJDK's java.base module, which doesn't work because
they depend on classes that belong to other modules on
host.

This CL fixes compilation to make the patch_module
attribute specific to the device target, like it
was before the aforementioned commit.

Bug: 115604102
Test: EXPERIMENTAL_USE_OPENJDK9=true make android-icu4j-tests
Change-Id: I5158e98c3dd0a9cc2b1e69ce553c27cce0aeae8a