Support adding default constructors and @UnsupportedAppUsage
AddDefaultConstructor will add a default constructor to a whitelist of
classes. It is needed because in some cases the @UnsupportedAppUsage
needs to be added to a default constructor that is not present in the
upstream source.

Refactored AddMarkerAnnotation to extract code that is common between
it and the new AddAnnotation class. The code for inserting the
annotation has been generalized to support all the possible types of
annotations based on information provided in an AnnotationInfo object.
The AnnotationInfo references an AnnotationClass which contains
information about the types of an annotation's properties which is
needed for parsing.

Added BodyDeclarationLocatorMap to provide a mapping from
BodyDeclaration to some object, in this case the AnnotationInfo to
insert.

Refactored code for converting String[] to List<TypeLocator> so that
it could be easily reused.

The locations and properties of the UnsupportedAppUsage annotations
that must be added are provided in a JSON file
(tools/srcgen/unsupported-app-usage.json) rather than in a flat file
as used for the CorePlatformApi. That is needed in future to add
properties, e.g. maxTargetSdk, to the UnsupportedAppUsage annotations.

This does not actually change the repackaged icu code, that will be
done in a follow up change.

Bug: 117818301
Test: make checkbuild, functionality tested in follow up change
Change-Id: I0a027c90c7c57251cab1a9f5939198f0a80f703c