Annotate repackaged API with @UnsupportedAppUsage
Adds @UnsupportedAppUsage annotations to the parts of ICU that are in
the hidden API. Allows the ICU entries to be removed from the
frameworks/base/config/hiddenapi-light-greylist.txt.

Also adds a few default constructors that are not present upstream but
which need to be added to the source so they can be annotated with
UnsupportedAppUsage.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Bug: 117818301
Test: make checkbuild and
Change-Id: I01d91810b3bbac4300aa91b992692bdc61476318