Add --default-constructors-file option to RepackagingTransform
Will allow projects that use RepackagingTransform to add default
constructors into the source during transformation. That is usually so
that they can be annotated and so the rule to do that runs before the
AddAnnotation rules.

Some minor cleanup:
In TypeLocator:
* Fixed loop condition that was always true.
* Added readTypeLocators(Path) to read a list from a file and
  readLines(Path) to read lines from a file, after filtering out blank
  lines and # comments.
* Cleaned up some unchecked warnings.

In BodyDeclarationLocators:
* Used TypeLocator.readLines(Path) in readBodyDeclarationLocators(Path).

In AddDefaultConstructor:
* Removed unused imports.

Tested by running scripts in all the projects that use this and
verifying that the generated files did not change.

Bug: 117818301
Test: see above
Change-Id: I2b50c47f0ecdde8c91bc4076bf509cf71990d2c8