Make AddAnnotation stricter in parsing JSON file
Using lenient mode allows too many types of comments which makes it
difficult to process reliably. This change filters out line comments
from the input JSON and then parses it in strict mode.

Tested by running scripts in all the projects that use this and
verifying that the generated files did not change.

Bug: 117818301
Test: see above

Change-Id: Icb09196c20b4afc8015cd96febd5280737ed9ade