Deprecate BreakIterator.getTitleInstance() functions and KIND_TITLE field
The following functions and constant have been deprecated since
Unicode 3.2 and are likely to be removed in ICU 64 (but not confirmed
oficially):

 * public static final int KIND_TITLE = 4;
 * public static BreakIterator getTitleInstance()
 * public static BreakIterator getTitleInstance(Locale where)
 * public static BreakIterator getTitleInstance(ULocale where)

This CL adds these functions and constant to ANDROID_DEPRECATED set in
Icu4jTransform.java effectively marking them as deprecated. It also
modifies javadoc for deprecated BreakIterator#KIND_TITLE fiels to
contain the reference to another BreakIterator instance that should be
used instead.

Bug: 123390510
Test: run generate_android_icu4j.sh, ensure m droid builds successfully
Change-Id: Ib2bd93e0d0467de1d46245cfd626a717a1dfac5b