Add a static library to share ICU init code
Add a static library (androidicuinit) to enable sharing of ICU4C
initialization code. Switch AIcu.cpp over to using it. This means
that libcore and various command line utils now initialize ICU4C
in the same way.

This means for binaries that use AIcu_initializeIcuOrDie():
1) They previously didn't use time zone data overlay files from
/data and/or the time zone data mainline module but they now do.
2) ICU4C is initialized in a different way than before: rather than
finding the files from a given path, the files are mapped into
memory and the memory locations are explicitly given to ICU4C.

Testing:
To check for the expected memory mapping:
$ pmap <pid> | grep '.dat'
with zygote and mediaserver process pids.

Tested sqlite3 after executing in adb shell as above (but a sql
operation like "CREATE TABLE" has to be attempted before ICU
is initialized).

Also inspected logcat output for the expected log messages.

Bug: 120493361
Test: build / boot
Test: atest CtsLibcoreTestCases
Test: See above
Change-Id: I7c102ddafaa36e1ff81f1dfbce4f779553ba71a1