Fix "Could not find jar: ICU4CTestRunner.jar"
ICU4CTestRunner.jar wasn't not saved into the same directory
as intltest and cintltst. Thus, when partial download happens in the
test infra, the .jar file is not found.

Fix this by creating a new build module including the native test
binary and ICU4CTestRunner.jar.

The downside is that the test does not run in 32-bit mode on 64-bit
device. It's fine as the ICU4C codes are mostly CPU architecture
independent.

Also, convert the test from general-tests to device-tests because
the ABIs being used may not be stable, and may depend on compiler
flags or CPU architecture.

Bug: 141637587
Test: atest intltest_test
Test: atest cintltst_test
Change-Id: I1037b1495d1811a470f889818bd0d0976d189cf6