silence -Wunreachable-code-loop-increment
This warning is going to be enabled and become an error. It fires here:
external/icu/icu4c/source/i18n/numparse_compositions.cpp:67 . A comment
above the loop clarifies that the loop is "meant to be used for boundary
checking," and is only meant to run up to once.

Bug: 150166387
Test: TreeHugger
Change-Id: Ic8ff81da4d70ab807ff310ea701b077b738c624f