Add windows x86_64 libicuuc to artifacts for layoutlib
Layoutlib on Windows links libicuuc (and transitively on
libiccuc_stubdata) dynamically. We need to add those libraries to the
list of build artifacts for layoutlib so that they can be downloaded
from ab/ and added to Android Studio.
We only need the x86_64 version as Android Studio does not support
32-bit Windows any more.

Bug: 149286037
Test: m layoutlib dist
Change-Id: I6a5aedbebc848485985cf498164c1b5a1df3a8c6