Mark libandroidicu_headers as available to platform
The header lib is depended on by libpdfium-fxcrt that is part of the
platform. (the dependency was added recently with aosp/1296928)

With b/153073816, we will be checking the availability to the platform,
i.e. platform-available module depending on platform-unavailable module
causes a build error.

Prepare for the change by correctly marking that this interface lib is
available to the platform.

Bug: 153073816
Test: m
Change-Id: Ic08235454603a5a5eae0f9311b753231275ddbab