Android patch: Fix localized quarters used in java.time
Resource alias wasn't handled correctly in the old implementation
in DateTimeTextProvider, for example, the wide standalone quarter
for Q1 in the locale *agq* should be localized to "kɨbâ kɨ 1",
instead of "Q1" provided by the ROOT resources.

And it replaces the ICUResourceBundle usage with the new APIs,
because ICUResourceBundle should not exposed as @IntraCoreApi.

Bug: 139480281
Bug: 170717042
Test: ExtendedDateFormatSymbolsTest
Test: atest CtsLibcoreOjTestCases
Change-Id: I250d391163c6c7845a0a284141287d3b69d654cf