Remove u_strCompareIter and uloc_acceptLanguageFromHTTP from NDK headers
It does not cause any compat problems because the functions are not
the NDK stub described in .map.txt. Also, there were no new release
after adding the NDK headers, and no app can use them today.

It was accidentally added into the headers due to a bug in the
code generation script.

Bug: 160350521
Test: m droid cts
Change-Id: I53d8984df6ce20578ccabb4572ab0f3586125d27