Rename and document @IntraCoreApi in com.android
For renaming, I target using com.android.icu or com.android.i18n
package name for @IntraCoreApi only. The @IntraCoreApi in android.icu
(the package name for ICU4J public APIs) will be removed in the next
few CLs.

The goal is to move all @IntraCoreApi away from android.icu package
because the upstream may change the API signature in the future.
This CL creates a @IntraCoreApi classes in com.android.icu for
a subset of @IntraCoreApi in android.icu. Corresponding @IntraCoreApi
in android.icu will be removed when the user in libcore is updated.

This CL also documents all @IntraCoreApi in com.android.icu and
their API signatures are considered in the final state for the
Android S release.

Bug: 139480281
Test: CtsLibcoreTestCases
Change-Id: I8a6198d1aa327ff989593c882b71f94dfddffadd