Cherry-pick: Reapply ICU's upstream tzdb 2020b changes for ICU 67.
This contains a partial cherry-pick of upstream commit
0080c0c5e7948ab273d6bf6f341e1091afbf432d for
ICU-21325 tz database 2020b updates.
https://github.com/unicode-org/icu/commit/0080c0c5e7948ab273d6bf6f341e1091afbf432d

This is needed because ICU 67 was cut with TZDB 2019c and
Android has been updated to TZDB 2020b. The original version
of this change on master was commit 5f113a981ee767e1cb5393b989a5aecb13980dd0,
and this is effectively a reapply.

The .txt / .dat file updates from the original are not included. They will be
regenerated by a later commit in the chain for the ICU 67 update.

Bug: 170607822
Change-Id: Idf0df507cb5abde7722e0a00363a37e6bbac0868