Android patch: Add Calendar.getDateTimeFormatString used by java.time via libcore bridge
Add Calendar.getDateTimeFormatString needed by java.time.
It tends to expose functionality that most end user code
will not need, but allows access to ICU internal data or functionality
that should be shared between ICU and java.time.

- Add method to Calendar to get localized pattern string.

This change was introduced in Android for the O release:

https://android.googlesource.com/platform/external/icu/+/22b47ef

It was then amended to remove DateFormatSymbols(ULocale, String)
constructor in ICU 69, because no one is using it.

https://r.android.com/q/I1d65bca1c301b5e13e265062c9eeab1196edb516

Test: n/a
Change-Id: I1d65bca1c301b5e13e265062c9eeab1196edb516